Remove relying on "encoding/json"
This originally depended on "encoding/json" if no options are passed and
only depended on "github.com/go-json-experiment/v1" if options are passed.

This was a faulty attempt at making it possible to detect
divergences in behavior between the two, but it does not work.

We may need to add explicit API to detect regressions between
"encoding/json" and "github.com/go-json-experiment/json/v1".
However, this is complicated by the fact that they will actually
be the same package once Go 1.25 is released and
the binary is built with GOEXPERIMENT=jsonv2.