Use function-relative line numbering
The file:line notation is less useful when different versions
of a binary are being deployed as the lines become stale
relatively quickly with even minor modifications to the source file.

The function-relative line numbering is more stable,
which is why the Go telemetry service uses this notation.
See https://research.swtch.com/telemetry-design