Add ErrNotCloneable and report failures to clone (#12)
Codec.Unmarshal only works if the output value can be cloned.
However, if it cannot be cloned, then calling both v1 and v2
is not possible, and thus no differences will ever be detected.

This gives the false assurance that v1 and v2 are identical
since the differences count will be zero and users may forget
to check that the NumUnmarshalCallBothSkipped count is also zero.

Thus, treat failure to clone as a difference and
surface such cases by calling ReportDifference.
The user-provided ReportDifference function can detect such cases
since ErrorV1 or ErrorV2 will be set to the ErrNotCloneable error.